Fix: Minor fixes to enable compilation with clang-cl

+++ src/problems/AOC_2024/AOC_2024_day_25.cpp
@@ -51,6 +51,6 @@ size_t day25_1(std::string dataFile) {
         return cmp == 5;
     });
 
-    return res;
+    return static_cast<size_t>(res);
 }
 } // namespace AOC2024
